requirements.txt
numpy>=2.2.5
pydantic>=2.9.0
PyYAML>=6.0.1
pandas>=2.2.3
scipy>=1.15.2
scikit-learn>=1.6.1
matplotlib>=3.10.1
seaborn>=0.13.2
chembl-webresource-client>=0.10.9
Pillow>=11.2.1
lazypredict>=0.2.16
click>=8.1.8
lightgbm>=4.6.0
tqdm>=4.67.1
xgboost>=3.0.0
shap>=0.47.2


